Update plublish process to publish to sonatype (#7)
Description
===========

Bintray and with it JCenter will be shutdown the comming weeks.
To be able to host this library in a commenly available repository
we need to switch the publish logic to publish to sonatype and with
it the central repository. This needs some adjustments in both
gradle file and jenkins pipeline.

I took the liberty to upgrade the project to use the latest version
of gradle `6.8.2` as this has some features regarding pgp signing which
gradle 5 does not have. I want to sign the artifacts with a gpg subkey
to not roll out the master key on the ci system with a ascii in memory
format which is not supported with gradle 5.

The jenkins steps also need to fetch new credentials.

Changes
=======

* ![UPDATE] `build.gradle` to publish to OSSRH instead of bintray
* ![UPDATE] gradle wrapper to version `6.8.2`
* ![UDDATE] `Jenkinsfile` to pass new credentials and publish snapshot versions